refactor: モデル内のロジックをモジュールへ分割.
